Addressed some PR concerns
