+++ src/main/java/org/openstreetmap/atlas/locale/IsoCountry.java
@@ -1,33 +1,23 @@
 package org.openstreetmap.atlas.locale;
 
 import java.io.Serializable;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
-import java.util.HashMap;
-import java.util.List;
 import java.util.Locale;
 import java.util.Map;
-import java.util.Map.Entry;
 import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
-
-import org.apache.commons.lang3.StringUtils;
-import org.openstreetmap.atlas.exception.CoreException;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 /**
  * Java Locale based countries, including ISO2, ISO3, and descriptive country name representations.
  *
  * @author robert_stack
+ * @author lcram
  */
 public final class IsoCountry implements Serializable
 {
     private static final long serialVersionUID = 8686298246454085812L;
-
-    private static final Logger logger = LoggerFactory.getLogger(IsoCountry.class);
 
     // Use United States fixed Locale for display use cases
     private static final String LOCALE_LANGUAGE = Locale.ENGLISH.getLanguage();
@@ -35,14 +25,15 @@ public final class IsoCountry implements Serializable
     private static final int ISO2_LENGTH = 2;
     private static final int ISO3_LENGTH = 3;
 
-    private static final Set<String> ALL_COUNTRY_CODES;
-    private static final Set<String> ALL_DISPLAY_COUNTRIES;
-    private static final Map<String, String> ISO2_TO_DISPLAY_COUNTRY;
-    private static final Map<String, String> DISPLAY_COUNTRY_TO_ISO2;
-    private static final Map<String, String> ISO2_TO_ISO3;
-    private static final Map<String, String> ISO3_TO_ISO2;
+    // Package private fields used by other classes in the locale package
+    static final Set<String> ALL_COUNTRY_CODES;
+    static final Set<String> ALL_DISPLAY_COUNTRIES;
+    static final Map<String, String> ISO2_TO_DISPLAY_COUNTRY;
+    static final Map<String, String> DISPLAY_COUNTRY_TO_ISO2;
+    static final Map<String, String> ISO2_TO_ISO3;
+    static final Map<String, String> ISO3_TO_ISO2;
     // private static final BiMap<String, String> ISO2_ISO3_MAP;
-    private static final Map<String, IsoCountry> ISO_COUNTRIES;
+    static final Map<String, IsoCountry> ISO_COUNTRIES;
 
     static
     {
@@ -169,63 +160,14 @@ public final class IsoCountry implements Serializable
     }
 
     /**
-     * Provides IsoCountry for a valid country display name.
+     * Provides IsoCountry for a valid country display name. Ignores capitalization (e.g. "united
+     * stAtes" and "United States" are the same)
      *
      * @param displayCountry
      *            the display country name, e.g. "United States"
      * @return an Optional containing the IsoCountry if present
      */
     public static Optional<IsoCountry> forDisplayCountry(final String displayCountry)
-    {
-        if (displayCountry != null && DISPLAY_COUNTRY_TO_ISO2.containsKey(displayCountry))
-        {
-            return Optional
-                    .ofNullable(ISO_COUNTRIES.get(DISPLAY_COUNTRY_TO_ISO2.get(displayCountry)));
-        }
-        return Optional.empty();
-    }
-
-    /**
-     * Provides IsoCountry for a country display name. If the given display name does not perfectly
-     * match a valid IsoCountry, this will return the closest string match.
-     *
-     * @param displayCountry
-     *            the display country name, e.g. "united stats"
-     * @return an Optional containing the IsoCountry if present
-     */
-    public static Optional<IsoCountry> forDisplayCountryClosestMatch(final String displayCountry)
-    {
-        if (displayCountry != null)
-        {
-            if (DISPLAY_COUNTRY_TO_ISO2.containsKey(displayCountry))
-            {
-                return Optional
-                        .ofNullable(ISO_COUNTRIES.get(DISPLAY_COUNTRY_TO_ISO2.get(displayCountry)));
-            }
-            else
-            {
-                final Optional<String> closestCountry = closestIsoCountry(displayCountry);
-                if (closestCountry.isPresent())
-                {
-                    final Optional<IsoCountry> closestMatch = Optional.ofNullable(
-                            ISO_COUNTRIES.get(DISPLAY_COUNTRY_TO_ISO2.get(closestCountry.get())));
-                    logger.info("Exact match for {} was not found, returning closest match {}",
-                            displayCountry, closestMatch);
-                    return closestMatch;
-                }
-            }
-        }
-        return Optional.empty();
-    }
-
-    /**
-     * Provides IsoCountry for a valid country display name, ignoring case.
-     *
-     * @param displayCountry
-     *            the display country name, e.g. "united states"
-     * @return an Optional containing the IsoCountry if present
-     */
-    public static Optional<IsoCountry> forDisplayCountryIgnoreCase(final String displayCountry)
     {
         if (displayCountry != null)
         {
@@ -245,46 +187,6 @@ public final class IsoCountry implements Serializable
             return Optional.ofNullable(ISO_COUNTRIES.get(DISPLAY_COUNTRY_TO_ISO2.get(foundKey)));
         }
         return Optional.empty();
-    }
-
-    /**
-     * Provides closest IsoCountries for a country display name. If the given display name does not
-     * perfectly match a valid IsoCountry, this will return the closest string match up to number of
-     * matches.
-     *
-     * @param number
-     *            the number of matches to show
-     * @param displayCountry
-     *            the display country name, e.g. "united stats"
-     * @return an Optional containing the IsoCountry if present
-     */
-    public static List<IsoCountry> forDisplayCountryTopMatches(final int number,
-            final String displayCountry)
-    {
-        if (displayCountry != null)
-        {
-            final List<IsoCountry> results = new ArrayList<>();
-            if (DISPLAY_COUNTRY_TO_ISO2.containsKey(displayCountry))
-            {
-
-                results.add(ISO_COUNTRIES.get(DISPLAY_COUNTRY_TO_ISO2.get(displayCountry)));
-            }
-            else
-            {
-                final List<String> closestCountries = closestIsoCountries(number, displayCountry);
-                if (!closestCountries.isEmpty())
-                {
-                    logger.info("Exact match for {} was not found, returning closest {} matches {}",
-                            displayCountry, number, closestCountries);
-                    results.addAll(closestCountries.stream()
-                            .map(countryString -> ISO_COUNTRIES
-                                    .get(DISPLAY_COUNTRY_TO_ISO2.get(countryString)))
-                            .collect(Collectors.toList()));
-                }
-            }
-            return results;
-        }
-        return new ArrayList<>();
     }
 
     /**
@@ -343,43 +245,6 @@ public final class IsoCountry implements Serializable
             }
         }
         return false;
-    }
-
-    private static List<String> closestIsoCountries(final int number, final String displayCountry)
-    {
-        if (number <= 0 || number > ALL_DISPLAY_COUNTRIES.size())
-        {
-            throw new CoreException(
-                    "number " + number + " out of range (0, " + ALL_DISPLAY_COUNTRIES.size() + ")");
-        }
-        final Map<String, Integer> countryRankings = new HashMap<>();
-        for (final String countryName : ALL_DISPLAY_COUNTRIES)
-        {
-            final int distance = StringUtils.getLevenshteinDistance(displayCountry, countryName);
-            countryRankings.put(countryName, distance);
-        }
-        final List<Entry<String, Integer>> entries = new ArrayList<>(countryRankings.entrySet());
-        Collections.sort(entries,
-                (entry1, entry2) -> entry1.getValue().compareTo(entry2.getValue()));
-
-        return entries.subList(0, number).stream().map(Entry::getKey).collect(Collectors.toList());
-    }
-
-    private static Optional<String> closestIsoCountry(final String displayCountry)
-    {
-        String closestCountry = null;
-        int minimumDistance = Integer.MAX_VALUE;
-        for (final String countryName : ALL_DISPLAY_COUNTRIES)
-        {
-            final int distance = StringUtils.getLevenshteinDistance(displayCountry, countryName);
-            if (distance < minimumDistance)
-            {
-                closestCountry = countryName;
-                minimumDistance = distance;
-            }
-        }
-
-        return Optional.ofNullable(closestCountry);
     }
 
     private IsoCountry(final String iso2)
