+++ src/test/java/org/openstreetmap/atlas/locale/IsoCountryTest.java
@@ -1,11 +1,9 @@
 package org.openstreetmap.atlas.locale;
 
 import java.util.Optional;
-import java.util.stream.Collectors;
 
 import org.junit.Assert;
 import org.junit.Test;
-import org.openstreetmap.atlas.utilities.collections.Lists;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -70,24 +68,7 @@ public class IsoCountryTest
                 IsoCountry.forDisplayCountry("United States").get().getIso3CountryCode());
         Assert.assertFalse(IsoCountry.forDisplayCountry("united states").isPresent());
         Assert.assertFalse(IsoCountry.forDisplayCountry("ZZzz zzzzz").isPresent());
-        Assert.assertEquals("USA",
-                IsoCountry.forDisplayCountryIgnoreCase("united states").get().getIso3CountryCode());
-        Assert.assertEquals("USA",
-                IsoCountry.forDisplayCountryIgnoreCase("UnITeD STatEs").get().getIso3CountryCode());
-        Assert.assertEquals("USA", IsoCountry.forDisplayCountryClosestMatch("UnitedStates").get()
-                .getIso3CountryCode());
-        Assert.assertEquals("USA", IsoCountry.forDisplayCountryClosestMatch("united States").get()
-                .getIso3CountryCode());
-        Assert.assertEquals("USA", IsoCountry.forDisplayCountryClosestMatch("united stats").get()
-                .getIso3CountryCode());
-        Assert.assertEquals("GBR", IsoCountry.forDisplayCountryClosestMatch("unitde kindgdom").get()
-                .getIso3CountryCode());
-        Assert.assertFalse(IsoCountry.forDisplayCountryIgnoreCase("ZZzz zzzzz").isPresent());
 
         Assert.assertTrue(IsoCountry.allDisplayCountries().contains("United States"));
-
-        Assert.assertEquals(Lists.arrayList("Sweden", "Cambodia", "Yemen"),
-                IsoCountry.forDisplayCountryTopMatches(3, "abcdefg").stream()
-                        .map(IsoCountry::getDisplayCountry).collect(Collectors.toList()));
     }
 }
